+++ lib/money/exchange_rates/exchange_rates_retriever.ex
@@ -223,7 +223,7 @@ defmodule Money.ExchangeRates.Retriever do
   end
 
   defp process_response({:error, {:tls_alert, {:certificate_expired, _message}}}, url, _config) do
-    {:error, "Certificate for #{inspect url} has expired"}
+    {:error, "Certificate for #{inspect(url)} has expired"}
   end
 
   defp if_none_match_header(url) do
@@ -495,71 +495,73 @@ defmodule Money.ExchangeRates.Retriever do
   #### Certificate verification
 
   @certificate_locations [
-      # Configured cacertfile
-      Application.get_env(Cldr.Config.app_name(), :cacertfile),
+                           # Configured cacertfile
+                           Application.get_env(Cldr.Config.app_name(), :cacertfile),
 
-      # Populated if hex package CAStore is configured
-      if(Code.ensure_loaded?(CAStore), do: CAStore.file_path()),
+                           # Populated if hex package CAStore is configured
+                           if(Code.ensure_loaded?(CAStore), do: CAStore.file_path()),
 
-      # Populated if hex package certfi is configured
-      if(Code.ensure_loaded?(:certifi), do: :certifi.cacertfile() |> List.to_string),
+                           # Populated if hex package certfi is configured
+                           if(Code.ensure_loaded?(:certifi),
+                             do: :certifi.cacertfile() |> List.to_string()
+                           ),
 
-      # Debian/Ubuntu/Gentoo etc.
-      "/etc/ssl/certs/ca-certificates.crt",
+                           # Debian/Ubuntu/Gentoo etc.
+                           "/etc/ssl/certs/ca-certificates.crt",
 
-      # Fedora/RHEL 6
-      "/etc/pki/tls/certs/ca-bundle.crt",
+                           # Fedora/RHEL 6
+                           "/etc/pki/tls/certs/ca-bundle.crt",
 
-      # OpenSUSE
-      "/etc/ssl/ca-bundle.pem",
+                           # OpenSUSE
+                           "/etc/ssl/ca-bundle.pem",
 
-      # OpenELEC
-      "/etc/pki/tls/cacert.pem",
+                           # OpenELEC
+                           "/etc/pki/tls/cacert.pem",
 
-      # CentOS/RHEL 7
-      "/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem",
+                           # CentOS/RHEL 7
+                           "/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem",
 
-      # Open SSL on MacOS
-      "/usr/local/etc/openssl/cert.pem",
+                           # Open SSL on MacOS
+                           "/usr/local/etc/openssl/cert.pem",
 
-      # MacOS & Alpine Linux
-      "/etc/ssl/cert.pem"
-  ]
-  |> Enum.reject(&is_nil/1)
+                           # MacOS & Alpine Linux
+                           "/etc/ssl/cert.pem"
+                         ]
+                         |> Enum.reject(&is_nil/1)
 
   def certificate_store do
     @certificate_locations
     |> Enum.find(&File.exists?/1)
     |> raise_if_no_cacertfile
-    |> :erlang.binary_to_list
+    |> :erlang.binary_to_list()
   end
 
   defp raise_if_no_cacertfile(nil) do
     raise RuntimeError, """
-      No certificate trust store was found.
-      Tried looking for: #{inspect @certificate_locations}
+    No certificate trust store was found.
+    Tried looking for: #{inspect(@certificate_locations)}
 
-      A certificate trust store is required in
-      order to download locales for your configuration.
+    A certificate trust store is required in
+    order to download locales for your configuration.
 
-      Since ex_cldr could not detect a system
-      installed certificate trust store one of the
-      following actions may be taken:
+    Since ex_cldr could not detect a system
+    installed certificate trust store one of the
+    following actions may be taken:
 
-      1. Install the hex package `castore`. It will
-         be automatically detected after recompilation.
+    1. Install the hex package `castore`. It will
+       be automatically detected after recompilation.
 
-      2. Install the hex package `certifi`. It will
-         be automatically detected after recomilation.
+    2. Install the hex package `certifi`. It will
+       be automatically detected after recomilation.
 
-      3. Specify the location of a certificate trust store
-         by configuring it in `config.exs`:
+    3. Specify the location of a certificate trust store
+       by configuring it in `config.exs`:
 
-         config :ex_cldr,
-           cacertfile: "/path/to/cacertfile",
-           ...
+       config :ex_cldr,
+         cacertfile: "/path/to/cacertfile",
+         ...
 
-      """
+    """
   end
 
   defp raise_if_no_cacertfile(file) do
@@ -568,8 +570,7 @@ defmodule Money.ExchangeRates.Retriever do
 
   defp https_opts(%Money.ExchangeRates.Config{verify_peer: true}) do
     [
-      ssl:
-      [
+      ssl: [
         verify: :verify_peer,
         cacertfile: certificate_store(),
         depth: 99,