+++ lib/money/parser/combinators.ex
@@ -4,7 +4,7 @@ defmodule Money.Combinators do
   import NimbleParsec
 
   # Whitespace as defined by Unicode set :Zs plus tab
-  @whitespace [?\s, ?\t, 0xa0, 0x1680, 0x2000, 0x202F, 0x205f, 0x3000]
+  @whitespace [?\s, ?\t, 0xA0, 0x1680, 0x2000, 0x202F, 0x205F, 0x3000]
   def whitespace do
     repeat(utf8_char(@whitespace))
     |> label("whitespace")
@@ -119,5 +119,4 @@ defmodule Money.Combinators do
     ])
     |> map(:change_sign)
   end
-
 end