+++ test/support/split_generator.ex
@@ -2,8 +2,10 @@ defmodule GenerateSplits do
   require ExUnitProperties
 
   def generate_money do
-    ExUnitProperties.gen all value <- StreamData.float(min: 0.0, max: 999_999_999_999_999.9),
-                             split <- StreamData.integer(1..101) do
+    ExUnitProperties.gen all(
+                           value <- StreamData.float(min: 0.0, max: 999_999_999_999_999.9),
+                           split <- StreamData.integer(1..101)
+                         ) do
       {Money.new(:USD, Float.to_string(value)), split}
     end
   end